+++ app/assets/javascripts/discourse/app/components/header/auth-buttons.gjs
