+++ app/assets/javascripts/discourse/app/components/header/contents.gjs
