+++ app/assets/javascripts/discourse/app/components/header/dropdown.gjs
