+++ app/assets/javascripts/discourse/app/components/header/hamburger-dropdown-wrapper.gjs
