+++ app/assets/javascripts/discourse/app/components/header/home-logo.gjs
