+++ app/assets/javascripts/discourse/app/components/header/icons.gjs
