+++ app/assets/javascripts/discourse/app/components/header/logo.gjs
