+++ app/assets/javascripts/discourse/app/components/header/panel-portal.gjs
