+++ app/assets/javascripts/discourse/app/components/header/search-menu-wrapper.gjs
