+++ app/assets/javascripts/discourse/app/components/header/sidebar-toggle.gjs
