+++ app/assets/javascripts/discourse/app/components/header/topic/featured-link.gjs
