+++ app/assets/javascripts/discourse/app/components/header/topic/info.gjs
