+++ app/assets/javascripts/discourse/app/components/header/topic/participant.gjs
