+++ app/assets/javascripts/discourse/app/components/header/topic/status.gjs
