+++ app/assets/javascripts/discourse/app/components/header/user-dropdown/notifications.gjs
