+++ app/assets/javascripts/discourse/app/components/header/user-dropdown/user-status-bubble.gjs
