+++ app/assets/javascripts/discourse/app/components/header/user-menu-wrapper.gjs
