+++ app/assets/javascripts/discourse/app/widgets/header.js
@@ -2,8 +2,8 @@ import { schedule } from "@ember/runloop";
 import { hbs } from "ember-cli-htmlbars";
 import $ from "jquery";
 import { h } from "virtual-dom";
-import { headerButtonsDAG } from "discourse/components/glimmer-header";
-import { headerIconsDAG } from "discourse/components/glimmer-header/icons";
+import { headerButtonsDAG } from "discourse/components/header";
+import { headerIconsDAG } from "discourse/components/header/icons";
 import { addExtraUserClasses } from "discourse/helpers/user-avatar";
 import { wantsNewWindow } from "discourse/lib/intercept-click";
 import scrollLock from "discourse/lib/scroll-lock";
