+++ app/assets/javascripts/discourse/app/widgets/home-logo.js
@@ -1,4 +1,4 @@
-// deprecated in favor of components/glimmer-header/home-logo.gjs
+// deprecated in favor of components/header/home-logo.gjs
 import { h } from "virtual-dom";
 import { wantsNewWindow } from "discourse/lib/intercept-click";
 import DiscourseURL from "discourse/lib/url";
