+++ app/assets/javascripts/discourse/app/widgets/sidebar-toggle.js
@@ -1,4 +1,4 @@
-// Deprecated in favor of app/assets/javascripts/discourse/app/components/glimmer-header/sidebar-toggle.gjs
+// Deprecated in favor of app/assets/javascripts/discourse/app/components/header/sidebar-toggle.gjs
 import { createWidget } from "discourse/widgets/widget";
 
 export default createWidget("sidebar-toggle", {
